project and project site apis completed and department updated

+++ src/main/java/com/grapplesoft/meil_backend/models/request/transactions/AllotProjectSiteRequestDto.java
@@ -38,7 +38,7 @@ public record AllotProjectSiteRequestDto(
         @JsonProperty("courierPmobile")
         String courierPmobile,
         @JsonProperty("addressId")
-        Long addressId,
+        int addressId,
         @JsonProperty("latitude")
         BigDecimal latitude,
         @JsonProperty("longitude")
